requirements.txt
# Dash Versions
dash-core-components==1.10.0
dash-daq==0.4.0
dash-html-components==1.0.3
dash-table==4.7.0
crystal-toolkit==2020.9.3
dash-mp-components==0.0.21
scikit-image==0.17.2
scikit-learn==0.23.2

# Data access
# monty>=3.0.4
# maggma>=0.22.3
